new_agent/requirements.txt
# Enhanced Agent Framework Requirements - LangGraph Integration

# Core dependencies for LLM-powered agentic operations
openai>=1.3.0
tiktoken>=0.5.2
rich>=13.0.0
pydantic>=2.0.0

# LangGraph for state management and streaming
langgraph>=0.2.0
langchain-core>=0.3.0
langchain-openai>=0.2.0

# Optional environment file support
python-dotenv>=1.0.0

# For development/testing (optional)
# pytest>=7.0.0
# pytest-asyncio>=0.21.0

# Note: This framework now uses OpenAI GPT-4o Mini for all agentic operations
# including planning, agent decisions, routing, and summarization.
# Only tool outputs for external APIs use dummy implementations.
# 
# Required Environment Variable:
# OPENAI_API_KEY=your_openai_api_key_here
# 
# You can also create a .env file in the project root with:
# OPENAI_API_KEY=your_openai_api_key_here